#2 Started writing a construct that allows for accessing java Methods on arbitrary objects.

+++ src/main/java/de/yanwittmann/menter/interpreter/structure/EvaluationContext.java
@@ -1288,7 +1288,7 @@ public abstract class EvaluationContext {
             s1 = s1.toLowerCase();
             s2 = s2.toLowerCase();
 
-            int score = 0;
+            double score = 0;
             int currentMatchIndex = 0;
             boolean isMatchingSequence = true;
             for (int i = 0; i < s1.length() && currentMatchIndex < s2.length(); i++) {
@@ -1305,7 +1305,7 @@ public abstract class EvaluationContext {
                 }
             }
 
-            return score;
+            return (int) score;
         };
 
         return stringCandidates.stream()
